requirement.txt
absl-py==1.2.0
asn1crypto==1.5.1
astunparse==1.6.3
attrdict==2.0.1
attrs==22.1.0
beautifulsoup4==4.11.1
bitarray==1.2.2
blake2b-py==0.1.4
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854175163/work
bs4==0.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1625835293160/work
cfgv==3.3.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1655906222726/work
coincurve==17.0.0
coloredlogs==15.0.1
commonmark==0.9.1
configparser==5.2.0
coverage==6.4.3
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967113783/work
crytic-compile==0.2.3
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==3.0.0a11
cytoolz==0.11.2
dictionaries==0.0.2
distlib==0.3.5
docopt==0.6.2
eth-abi==1.3.0
eth-account==0.3.0
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
ethereum-input-decoder==0.2.2
filelock==3.8.0
flatbuffers==1.12
fonttools==4.25.0
gast==0.4.0
gensim==4.2.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
hexbytes==0.2.3
humanfriendly==10.0
identify==2.5.3
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.1.2
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
libclang==14.0.6
loguru==0.6.0
lru-dict==1.1.8
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mock==4.0.3
munkres==1.1.4
mypy-extensions==0.4.3
mythril==0.23.6
networkx==2.8.5
nodeenv==1.7.0
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
oauthlib==3.2.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opt-einsum==3.3.0
oyente==0.2.7
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
parsimonious==0.8.1
persistent==4.9.0
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1621287973719/work
pip-search==0.0.12
pipreqs==0.4.11
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
prettytable==3.3.0
protobuf==3.19.4
psutil==5.7.2
py==1.11.0
py-ecc==1.7.1
py-evm==0.3.0a1
py-flags==1.1.4
py-solc==3.2.0
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pycryptodome==3.15.0
pyethash==0.1.27
Pygments==2.12.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pysha3==1.0.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857275402/work
pytest==7.1.2
pytest-cov==3.0.0
pytest-mock==3.8.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
PyYAML==6.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
requests-oauthlib==1.3.1
rich==12.5.1
rlp==1.2.0
rsa==4.9
rusty-rlp==0.1.15
scikit-learn @ file:///tmp/abs_d76175bc-917a-47d4-9994-b56265948a6328vmoe2o/croots/recipe/scikit-learn_1658419412415/work
scipy @ file:///tmp/build/80754af9/scipy_1641555001653/work
semantic-version==2.10.0
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
slither-analyzer==0.8.3
smart-open==6.0.0
solc-select==0.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchtext==0.13.1
torchvision==0.13.1
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1650891076910/work
transaction==3.0.1
trie==1.4.0
typing-extensions==3.10.0.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
virtualenv==20.16.3
wcwidth==0.2.5
web3==4.8.3
websockets==6.0
Werkzeug==2.2.2
wrapt==1.14.1
yarg==0.1.9
z3-solver==4.10.2.0
zipp==3.8.1
zope.interface==5.5.0.dev0
